Refactor: Use models from flowbot-commons, fix PR review comments
Addresses PR feedback:
1. ✅ Use schema-specific PyCRMBaseModel (not generic BaseModel)
2. ✅ Models now imported from flowbot-commons
3. ✅ All numeric fields use Decimal instead of float
4. ⏳ Address model - being handled separately
5. ✅ WarehouseMember uses WarehouseMemberRole enum (removed role_name)
6. ✅ Update methods now use repository.update() pattern

Changes:
- Deleted local model files (now in commons)
- Updated all imports to use commons.db.v6
- Updated Strawberry types for Decimal and enum support
- Fixed service update methods to use base repository pattern

+++ app/graphql/v2/core/container_types/strawberry/container_type_input.py
@@ -1,9 +1,12 @@
 """Strawberry input types for container types."""
 
+from decimal import Decimal
+
 import strawberry
 
+from commons.db.v6 import ContainerType
+
 from app.core.strawberry.inputs import BaseInputGQL
-from app.graphql.v2.core.container_types.models import ContainerType
 
 
 @strawberry.input
@@ -11,18 +14,18 @@ class ContainerTypeInput(BaseInputGQL[ContainerType]):
     """Input type for creating/updating container types."""
 
     name: str
-    length: float  # in inches
-    width: float  # in inches
-    height: float  # in inches
-    weight: float  # tare weight in lbs
+    length: Decimal  # in inches
+    width: Decimal  # in inches
+    height: Decimal  # in inches
+    weight: Decimal  # tare weight in lbs
     order: int | None = None  # display order (auto-assigned if not provided)
 
     def to_orm_model(self) -> ContainerType:
         return ContainerType(
             name=self.name,
-            length=self.length,  # type: ignore[arg-type]
-            width=self.width,  # type: ignore[arg-type]
-            height=self.height,  # type: ignore[arg-type]
-            weight=self.weight,  # type: ignore[arg-type]
+            length=self.length,
+            width=self.width,
+            height=self.height,
+            weight=self.weight,
             order=self.order or 0,  # Will be auto-assigned if 0
         )
